+++ scripts/jetstream/cron_end_date_volumes_invariant_13a.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+HOME="/opt/dev/atmosphere"
+VIRTUAL="/opt/env/atmo"
+export DJANGO_SETTINGS_MODULE="atmosphere.settings"
+export PYTHONPATH="$HOME:$PYTHONPATH"
+cd $HOME
+. $VIRTUAL/bin/activate
+echo $PATH
+echo $PYTHONPATH
+$HOME/scripts/end_date_volumes_invariant_13a.py