+++ CHANGELOG.md
@@ -28,6 +28,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
   - Add support for multi-instance-launch, 1 request can launch multiple instances by passing `instance_count` as an extra parameter
     ([#736](https://github.com/cyverse/atmosphere/pull/736))
+    
+### Changed
+  - Update and add Jetstream scripts
+  ([#735](https://github.com/cyverse/atmosphere/pull/735))
+  
 
 ## [v36-6](https://github.com/cyverse/atmosphere/compare/v36-5...v36-6) - 2019-10-22
 ### Changed